requirements.txt
absl-py==2.2.1
alembic==1.15.1
astunparse==1.6.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.6
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.71.0
h5py==3.13.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.2
joblib==1.3.1
keras==3.9.1
libclang==18.1.1
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.0.8
numpy==1.26.4
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
Pillow==10.0.0
protobuf==4.25.6
Pygments==2.19.1
PyMySQL==1.1.1
python-dateutil==2.8.2
python-dotenv==1.1.0
pytz==2023.3
requests==2.32.3
rich==14.0.0
scikit-learn==1.6.1
scipy==1.15.2
setuptools==78.1.0
six==1.16.0
SQLAlchemy==2.0.39
tensorboard-data-server==0.7.2
termcolor==2.5.0
threadpoolctl==3.2.0
typing_extensions==4.13.0
tzdata==2023.3
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
WTForms==3.0.1
